# Return of the threaded detection