--- a/src/ProgressMeter.jl
+++ b/src/ProgressMeter.jl
@@ -81,7 +81,8 @@ Base.@kwdef mutable struct ProgressCore
     numprintedvalues::Int       = 0             # num values printed below progress in last iteration
     prev_update_count::Int      = 1             # counter at last update
     printed::Bool               = false         # true if we have issued at least one status update
-    safe_lock::Bool             = Threads.nthreads() > 1 # set to false for non-threaded tight loops
+    safe_lock::Int              = 2*(Threads.nthreads()>1) # 0: no lock, 1: lock, 2: detect
+    thread_id::Int              = Threads.threadid() # id of the thread that created the progressmeter
     tinit::Float64              = time()        # time meter was initialized
     tlast::Float64              = time()        # time of last update
     tsecond::Float64            = time()        # ignore the first loop given usually uncharacteristically slow
@@ -448,8 +449,20 @@ end
 
 predicted_updates_per_dt_have_passed(p::AbstractProgress) = p.counter - p.prev_update_count >= p.check_iterations
 
+function is_threading(p::AbstractProgress)
+    p.safe_lock == 0 && return false
+    p.safe_lock == 1 && return true
+    if p.thread_id != Threads.threadid()
+        lock(p.lock) do
+            p.safe_lock = 1
+        end
+        return true
+    end
+    return false
+end
+
 function lock_if_threading(f::Function, p::AbstractProgress)
-    if p.safe_lock
+    if is_threading(p)
         lock(p.lock) do
             f()
         end